refactor: moved route types to shared according FSD ruleset
