+++ src/app/router/AppRoutes.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
+import { ROUTES } from '@/shared/lib/routes';
 import { RouterView } from 'vue-router';
 import { useRouter } from 'vue-router';
-import { ROUTES } from '.';
 
 const router = useRouter();
 const isAuth = true; // TODO: implement auth
