+++ src/shared/lib/routes.ts
@@ -1,0 +1,8 @@
+export const ROUTES = {
+  HOME: '/home',
+  LOGIN: '/login',
+  SIGNUP: '/signup',
+  NOT_FOUND: '/:pathMatch(.*)*'
+} as const;
+
+export type RoutePaths = (typeof ROUTES)[keyof typeof ROUTES];
